Added spiral enemy

+++ include/entities/Bullet.hpp
@@ -10,7 +10,7 @@
 
 class Bullet : public Sprite {
 	public:
-		Bullet(Scene* scene, int x, int y, int xVel, int yVel, int damage = 1, bool enemy = false);
+		Bullet(Scene* scene, int x, int y, double xVel, double yVel, int damage = 1, bool enemy = false);
 		~Bullet();
 		void update(double delta);
 		int getDamage();
