+++ src/entities/Bullet.cpp
@@ -2,7 +2,7 @@
 #include "engine/Collision.hpp"
 #include "engine/AssetLoader.hpp"
 
-Bullet::Bullet(Scene* scene, int x, int y, int xVel, int yVel, int damage, bool enemy) : 
+Bullet::Bullet(Scene* scene, int x, int y, double xVel, double yVel, int damage, bool enemy) : 
     Sprite({
         enemy ?
         Sprite{AssetLoader::tilesheet, AssetLoader::enemyBullet, 4, .2, 0, 10, 10} :
